Enable compilation with stable Rust
To the Rust teams, Rust's contributors,
the entire Rust and Rocket communities,
and Sergio Benitez: thank you all. Sincerely.

Most of the code generation code is **heavily**
based on Rocket (https://rocket.rs/).
Also a lot of the history is shared with it,
including this very commit message.
So now that Rocket compiles on stable Rust
there was no reason that cuke-runner couldn't follow suit.

Thank you all!

+++ core/lib/src/api/location.rs
@@ -4,7 +4,7 @@ use std::path::{Path, PathBuf};
 #[derive(Debug, Clone)]
 pub struct GlueCodeLocation {
     pub(crate) file_path: PathBuf,
-    pub(crate) line_number: usize,
+    pub(crate) line_number: u32,
 }
 
 impl GlueCodeLocation {
@@ -14,7 +14,7 @@ impl GlueCodeLocation {
     }
 
     /// Returns the line number in the glue code file.
-    pub fn line_number(&self) -> usize {
+    pub fn line_number(&self) -> u32 {
         self.line_number
     }
 }
